Cycle Run
This has been done as a part of the project Cycle Run in MIT Media Lab's
Design and Innovation workshop-2015, GandhiNagar. A project where
physical aspect of cycling is combined with digital game to give an
engaging experience. Code for the game made using construct 2. Code for
Tessel which includes the accelerometer module and also it intergrates
with a web server hosted on Heroku, a cloud platform compatible with
Node.JS